vestigia/requirements.txt
# =============================================================================
# Vestigia - AI Agent Forensic Audit System
# Production Dependencies
# =============================================================================

# --- Core ---
python-dotenv==1.0.0

# --- API Server (Phase 3) ---
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
httpx==0.26.0

# --- Database (Phase 1/3) ---
psycopg2-binary==2.9.9

# --- Dashboard ---
streamlit==1.29.0
plotly==5.18.0
pandas==2.2.3

# --- Caching ---
redis==5.0.1

# --- File Monitoring (Watchtower) ---
watchdog==3.0.0

# --- OpenTelemetry (Phase 2) ---
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-prometheus==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0

# --- SIEM Integration (Phase 2) ---
# httpx already listed above (used for Splunk HEC, Elasticsearch, Datadog)

# --- Testing ---
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# --- Security ---
cryptography==41.0.7
PyJWT==2.8.0

# --- Playbooks ---
PyYAML==6.0.1
